Added database for Notes, Tasks and Machine service as well CropActivitities List

+++ app/src/main/java/com/myfarmnow/myfarmcrop/models/CropFertilizer.java
@@ -102,6 +102,6 @@ public class CropFertilizer implements Serializable,CropSpinnerItem {
         this.id = id;
     }
     public String getComposition(){
-        return "("+getnPercentage() +": "+getpPercentage()+" : "+getkPercentage()+" )";
+        return "("+getnPercentage() +": "+(int)getpPercentage()+" : "+getkPercentage()+" )";
     }
 }
